+++ src/records/UserLoginRecord.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace bymayo\porter\records;
+
+use Craft;
+use craft\db\ActiveRecord;
+
+class UserLoginRecord extends ActiveRecord
+{
+    public static function tableName()
+    {
+        return '{{%porter_user_logins}}';
+    }
+}